pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "modular-rag-mcp-server"
version = "0.1.0"
description = "A modular, pluggable RAG (Retrieval-Augmented Generation) service framework with MCP protocol support"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["rag", "mcp", "llm", "retrieval", "embedding"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core dependencies will be added in later stages
    "pytest>=9.0.2",
    "pyyaml>=6.0",
    "langchain>=0.1.0",  # B7.5: Recursive Splitter
    "langchain-text-splitters>=0.1.0",  # B7.5: RecursiveCharacterTextSplitter
    "chromadb>=0.4.0",  # B7.6: Chroma VectorStore
    "qdrant-client>=1.7.0",  # Qdrant VectorStore（支持 list 等复杂 metadata）
    "markitdown[pdf]>=0.0.1a26",  # C3: PDF Loader (需要 PDF 支持)
    "pymupdf>=1.23.0",  # 图片提取：从 PDF 中提取图片数据和位置信息
    "mcp[cli]>=1.0.0",  # 官方 MCP SDK (FastMCP + Stdio/HTTP)
    "jieba>=0.42.1",  # 中英文分词（SparseEncoder / BM25 检索）
    "sqlite-vec>=0.1.0",  # SQLite 向量扩展，用于 chunks_vec 虚拟表
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
dashboard = [
    "streamlit>=1.30.0",  # F4: Observability Dashboard
]
reranker = [
    "sentence-transformers>=2.2.0",  # B7.8: Cross-Encoder Reranker
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
